requirements.txt
openai==2.30.0
pytest==9.0.2
rich>=13.0
questionary>=2.0

# Web UI
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3

# Database ORM
sqlalchemy==2.0.25
psycopg2-binary==2.9.10

# Git Sync
gitpython==3.1.41

# Async testing
pytest-asyncio==0.23.3

# Environment variables
python-dotenv==1.0.0